Edits for File 5
For some reason certain things got deleted and input file 1 was not working. Fixed. 1 works fine now. Can't remember what was wrong with it, but some of the expects were missing the character in double quotes such as expect(")"); was expect(""); for some reason. Also edited and added functions to have 5 working. Input file 5 is erroring out at the correct location now but we need to add an actual error code for it as right now it is erroring out because: "There is a logic error in function Tokenizer::getToken.
The function doesn't know how to process this character: ->
<-"