Input 1,2,5 working
Fixed code so that 1, 2, and 5 work. For 2 the last 3 r braces are on the same line not sure if that should be the case. For 5 for some reason not all of my edits were merged so i had to redo the edits. It works now but still not correct error output.